Refactored feature steps
